% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findAxes.r
\name{findAxes}
\alias{findAxes}
\title{findAxes}
\usage{
findAxes(graph, func = walktrap.community, ...)
}
\arguments{
\item{graph}{An igraph object}

\item{func}{A function that returns a cluster membership for each node}

\item{\dots}{Further arguments for \code{func}}
}
\value{
A graph with a new Vertex slot 'axis' that stores the cluster membership
}
\description{
Places vertices into axes
}
\details{
Places vertices into axes based on a clustering algorithm
}
\examples{
data(flights)
g <- graph.data.frame(flights, directed=TRUE)
g2 <- findAxes(g)
V(g2)$axis
}
\author{
Jared P. Lander
}

