% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{flights}
\alias{flights}
\title{flights}
\format{A data.frame with three columns: \code{Origin}, \code{Destination} and \code{Price}.}
\usage{
flights
}
\description{
Edgelist, stored as a data.frame, containing information about flight paths.
}
\keyword{datasets}

