% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/layoutHive.r
\name{getAxisAngles}
\alias{getAxisAngles}
\title{getAxisAngles}
\usage{
getAxisAngles(graph, firstAxis = 0, degrees = FALSE)
}
\arguments{
\item{graph}{An igraph object}

\item{firstAxis}{By defualt the first axis is positioned along the x-axis.  This arugment allows that to be adjusted, which would offset all the axes by that amount.}

\item{degrees}{Logical indicating of firstAxis is specified in degrees.  Default is \code{FALSE}.}
}
\value{
A numerical vector giving the axis angle for each node in \code{graph}
}
\description{
Given a graph with vertice positions along axis, calculate the angles of the axes.
}
\details{
Given a graph with vertice positions along axis, calculate the angles of the axes.
}
\examples{
library(igraph)
data(flights)
g <- graph.data.frame(flights, directed=FALSE)
g <- findAxes(g)
getAxisAngles(g)
}
\author{
Jared P. Lander
}

