% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/layoutHive.r
\name{layout_with_hive}
\alias{layout_with_hive}
\title{layout_with_hive}
\usage{
layout_with_hive(graph)
}
\arguments{
\item{graph}{An igraph object}
}
\value{
A two-dimensional matrix with x- and y-coordinates
}
\description{
igraph layout function for hive plots
}
\details{
igraph layout function for hive plots
}
\examples{
library(igraph)
library(useful)
data(flights)
g <- graph.data.frame(flights, directed=FALSE)
l <- layout_with_hive(g)
plot(g, layout=l)
}
\author{
Jared P. Lander
}

