% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nodePosition.r
\name{nodePosition}
\alias{nodePosition}
\title{nodePosition}
\usage{
nodePosition(graph, func = closeness, ...)
}
\arguments{
\item{graph}{An igraph object}

\item{func}{A function that returns a cluster membership for each node}

\item{\dots}{Further arguments for \code{func}}
}
\value{
A graph with a new Vertex slot 'axis' that stores the cluster membership
}
\description{
Computes the position of vertices along axes
}
\details{
Computes the position of vertices along axes based on some centrality algorithm such as \code{\link[igraph]{closeness}}
}
\examples{
data(flights)
library(igraph)
g <- graph.data.frame(flights, directed=TRUE)
g2 <- nodePosition(g)
V(g2)$nodePosition
}
\author{
Jared P. Lander
}

