% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nodeSize.r
\name{nodeSize}
\alias{nodeSize}
\title{nodeSize}
\usage{
nodeSize(graph, func = page.rank, ...)
}
\arguments{
\item{graph}{An igraph object}

\item{func}{A function that returns a cluster membership for each node}

\item{\dots}{Further arguments for \code{func}}
}
\value{
A graph with a new Vertex slot 'axis' that stores the cluster membership
}
\description{
Computes size of vertices
}
\details{
Computes size of vertices based on some importance algorithm such as \code{\link[igraph]{page.rank}}
}
\examples{
data(flights)
g <- graph.data.frame(flights, directed=TRUE)
g2 <- nodeSize(g)
V(g2)$nodeSize
}
\author{
Jared P. Lander
}

